requirements.txt
modal
torch
torchvision
diffusers
transformers
accelerate
pillow
opencv-python
numpy
cupy-cuda12x
cuml-cu12
scour
cairosvg
fastapi
uvicorn
kagglehub